Fix bugs

+++ Trinity.SSA/JSCallCS.cs
@@ -313,11 +313,27 @@ namespace SSA
         public void DeleteQRCodeImageFileTemp()
         {
             Session session = Session.Instance;
-            Trinity.BE.User user = (Trinity.BE.User)session[CommonConstants.SUPERVISEE];
-            string fileName = String.Format("{0}/Temp/{1}", CSCallJS.curDir, "QRCode_" + user.NRIC + ".png");
-            if (System.IO.File.Exists(fileName))
-                System.IO.File.Delete(fileName);
+            Trinity.BE.User currentUser = (Trinity.BE.User)session[CommonConstants.USER_LOGIN];
+            if (currentUser != null)
+            {
+                Trinity.BE.User supervisee = null;
+                if (currentUser.Role == EnumUserRoles.DutyOfficer)
+                {
+                    supervisee = (Trinity.BE.User)session[CommonConstants.SUPERVISEE];
+                }
+                else
+                {
+                    supervisee = currentUser;
+                }
+                if (supervisee != null)
+                {
+                    string fileName = String.Format("{0}/Temp/{1}", CSCallJS.curDir, "QRCode_" + supervisee.NRIC + ".png");
+                    if (System.IO.File.Exists(fileName))
+                        System.IO.File.Delete(fileName);
+                }
+            }
         }
+
         public void popupLoading(string content)
         {
             this._web.LoadPopupHtml("LoadingPopup.html", content);
@@ -329,9 +345,15 @@ namespace SSA
             {
                 // Update queue status is finished
                 Session session = Session.Instance;
-                Trinity.BE.User user = (Trinity.BE.User)session[CommonConstants.SUPERVISEE];
+                Trinity.BE.User currentUser = (Trinity.BE.User)session[CommonConstants.USER_LOGIN];
+                if (currentUser != null)
+                {
+                    // Check why current user is null
+                    this._web.RunScript("$('.status-text').css('color','#000').text('The current user is null');");
+                    return;
+                }
                 var dalQueue = new DAL_QueueNumber();
-                dalQueue.UpdateQueueStatusByUserId(user.UserId, EnumStations.SSA, EnumQueueStatuses.Finished, EnumStations.UHP, EnumQueueStatuses.Processing, "", EnumQueueOutcomeText.Processing);
+                dalQueue.UpdateQueueStatusByUserId(currentUser.UserId, EnumStations.SSA, EnumQueueStatuses.Finished, EnumStations.UHP, EnumQueueStatuses.Processing, "", EnumQueueOutcomeText.Processing);
 
                 this._web.LoadPageHtml("PrintingMUBAndTTLabels.html");
                 this._web.RunScript("$('#WaitingSection').hide();$('#CompletedSection').show(); ; ");
@@ -340,10 +362,10 @@ namespace SSA
 
                 DeleteQRCodeImageFileTemp();
 
-                new DAL_QueueDetails().RemoveQueueFromSSK(user.UserId);
+                new DAL_QueueDetails().RemoveQueueFromSSK(currentUser.UserId);
 
-                Trinity.SignalR.Client.SignalR.Instance.QueueFinished(user.UserId);
-                //LogOut();
+                Trinity.SignalR.Client.SignalR.Instance.QueueFinished(currentUser.UserId);
+                LogOut();
             }
             else
             {
